+++ src/cli/cli.ts
@@ -1,6 +1,6 @@
-import { Command, program } from "commander"
+import { program } from "commander"
 import { config } from "dotenv"
-import * as packwiz from 'packwiz-extras-lib'
+import * as packwiz from '../lib/packwiz.ts'
 import * as fs from 'node:fs/promises'
 import { stringifyTOML } from 'confbox'
 import spawnSync from "./spawnSync.ts"
@@ -10,60 +10,56 @@ async function refresh(packFile: packwiz.Path) {
     spawnSync('packwiz', ['refresh'], { cwd })
 }
 
-async function replace(results: Map<packwiz.Path, packwiz.Mod>) {
+async function apply(result:packwiz.Result) {
     const jobs = []
-    for (const [path, mod] of results.entries()) {
-        jobs.push(fs.rm(path))
-        jobs.push(fs.writeFile(path + '.pw.toml', stringifyTOML(mod)))
+    if (result.delete !== undefined) {
+        for (const path of result.delete) {
+            jobs.push(fs.rm(path))
+        }
     }
-    await Promise.all(jobs)
-}
-
-async function overwrite(results: Map<packwiz.Path, packwiz.Mod>) {
-    const jobs = []
-    for (const [path, mod] of results.entries()) {
-        jobs.push(fs.writeFile(path, stringifyTOML(mod)))
+    if (result.write!==undefined) {
+        for (const [path, data] of Object.entries(result.write)) {
+            jobs.push(fs.writeFile(path,stringifyTOML(data)))
+        }
     }
-    await Promise.all(jobs)
 }
 
 config({ 'quiet': true })
 
 program.name('packwiz-extras')
     .description('extra utilities for packwiz')
-    .option('--pack-file <string>', 'The modpack file to use (default "pack.toml")', 'pack.toml')
+    .option('--pack-file <string>', 'The modpack file to use', 'pack.toml')
 
 const curseforge = program.command('curseforge').alias('cf')
     .description('manage curseforge files')
 
-const cfKey = ['--api-key <string>',
+const cfKey:string[] = ['--api-key <string>',
     'Your curseforge api key. This is required by curseforge to access their API. ' +
-    'Can also be provided via environment variable or .env file as CF_API_KEY',
-    process.env.CF_API_KEY]
+    'Can also be provided via environment variable or .env file as CF_API_KEY',]
 
-if (cfKey[2]) {
-    curseforge.option(cfKey[0], cfKey[1], cfKey[2])
+if (process.env.CF_API_KEY) {
+    curseforge.option(cfKey[0], cfKey[1])
 } else {
-    curseforge.requiredOption(cfKey[0], cfKey[1], cfKey[2])
+    curseforge.requiredOption(cfKey[0], cfKey[1])
 }
 
 curseforge.command('detect')
     .description('detect and replace files availible on curseforge')
     .action(async function () {
-        const apiKey = curseforge.opts().apiKey
+        const apiKey = curseforge.opts().apiKey || process.env.CF_API_KEY
         const packFile = program.opts().packFile
         const result = await packwiz.cfDetect(packFile, apiKey)
-        await replace(result)
+        await apply(result)
         await refresh(packFile)
     })
 
 curseforge.command('urls')
     .description('cache curseforge download urls to speed up packwiz-installer')
     .action(async function () {
-        const apiKey = curseforge.opts().apiKey
+        const apiKey = curseforge.opts().apiKey || process.env.CF_API_KEY
         const packFile = program.opts().packFile
         const result = await packwiz.cfUrl(packFile, apiKey)
-        await overwrite(result)
+        await apply(result)
         await refresh(packFile)
     })
 
@@ -78,7 +74,7 @@ modrith.command('detect')
         const apiKey = modrith.opts().apiKey
         const packFile = program.opts().packFile
         const result = await packwiz.mrDetect(packFile, apiKey)
-        await replace(result)
+        await apply(result)
         await refresh(packFile)
     })
 
@@ -88,7 +84,7 @@ modrith.command('merge')
         const apiKey = modrith.opts().apiKey
         const packFile = program.opts().packFile
         const result = await packwiz.mrMerge(packFile, apiKey)
-        await overwrite(result)
+        await apply(result)
         await refresh(packFile)
     })
